Update overview-of-shared-computer-activation-for-office-365-proplus.md
Adding the change so that customers are aware that there can be issues when switching to SCA after the machine has obtained a single user license.   When changing to SCA after the machine has obtained a single user license, the machine will continue to use the single user license and when the single user license attempts renewal Office will see that we are in SCA mode and the single user license will not renew.